Some refactors and adjustements

+++ Assets/Scripts/Enemy/HotHead/Enemy_hotHead.cs
@@ -11,6 +11,6 @@ public class Enemy_hotHead : MonoBehaviour
         _enemy = GetComponent<Enemy_actor>();
         _enemy.healthController.healthPoints = 1;
         _enemy.touchDamage = 1;
-        _enemy.type = (int)Powers.Fire;
+        _enemy.type = Powers.Fire;
     }
 }
